Changed simpcg to use 16-bit dimensions

+++ src/simpcg/scg-buffer.c
@@ -11,10 +11,10 @@
 #include "simpcg.h"
 
 static void scg_print_cell(struct SCGCell cell);
-static uint8_t scg_color_code_to_ansi_fg(enum SCGColorCode color_code);
-static uint8_t scg_color_code_to_ansi_bg(enum SCGColorCode color_code);
+static uint16_t scg_color_code_to_ansi_fg(enum SCGColorCode color_code);
+static uint16_t scg_color_code_to_ansi_bg(enum SCGColorCode color_code);
 
-struct SCGBuffer *scg_buffer_create(uint8_t width, uint8_t height)
+struct SCGBuffer *scg_buffer_create(uint16_t width, uint16_t height)
 {
 	size_t cells_size = width * height;
 	struct SCGBuffer *buffer = ALLOC_FLEX_STRUCT(buffer, cells, cells_size);
@@ -30,42 +30,42 @@ void scg_buffer_destroy(struct SCGBuffer *buffer)
 	free(buffer);
 }
 
-inline void scg_buffer_set_ch(struct SCGBuffer *buffer, uint8_t col, uint8_t row, char ch)
+inline void scg_buffer_set_ch(struct SCGBuffer *buffer, uint16_t col, uint16_t row, char ch)
 {
 	buffer->cells[row * buffer->width + col].ch = ch;
 }
 
-inline char scg_buffer_get_ch(struct SCGBuffer *buffer, uint8_t col, uint8_t row)
+inline char scg_buffer_get_ch(struct SCGBuffer *buffer, uint16_t col, uint16_t row)
 {
 	return buffer->cells[row * buffer->width + col].ch;
 }
 
-inline void scg_buffer_set_fg_color(struct SCGBuffer *buffer, uint8_t col, uint8_t row, enum SCGColorCode fg_color)
+inline void scg_buffer_set_fg_color(struct SCGBuffer *buffer, uint16_t col, uint16_t row, enum SCGColorCode fg_color)
 {
 	buffer->cells[row * buffer->width + col].fg_color = fg_color;
 }
 
-inline enum SCGColorCode scg_buffer_get_fg_color(struct SCGBuffer *buffer, uint8_t col, uint8_t row)
+inline enum SCGColorCode scg_buffer_get_fg_color(struct SCGBuffer *buffer, uint16_t col, uint16_t row)
 {
 	return buffer->cells[row * buffer->width + col].fg_color;
 }
 
-inline void scg_buffer_set_bg_color(struct SCGBuffer *buffer, uint8_t col, uint8_t row, enum SCGColorCode bg_color)
+inline void scg_buffer_set_bg_color(struct SCGBuffer *buffer, uint16_t col, uint16_t row, enum SCGColorCode bg_color)
 {
 	buffer->cells[row * buffer->width + col].bg_color = bg_color;
 }
 
-inline enum SCGColorCode scg_buffer_get_bg_color(struct SCGBuffer *buffer, uint8_t col, uint8_t row)
+inline enum SCGColorCode scg_buffer_get_bg_color(struct SCGBuffer *buffer, uint16_t col, uint16_t row)
 {
 	return buffer->cells[row * buffer->width + col].bg_color;
 }
 
 void scg_buffer_fill_ch(struct SCGBuffer *buffer, char ch)
 {
-	uint8_t width = buffer->width;
-	uint8_t height = buffer->height;
-	for (uint8_t row = 0; row < height; row++) {
-		for (uint8_t col = 0; col < width; col++) {
+	uint16_t width = buffer->width;
+	uint16_t height = buffer->height;
+	for (uint16_t row = 0; row < height; row++) {
+		for (uint16_t col = 0; col < width; col++) {
 			scg_buffer_set_ch(buffer, col, row, ch);
 		}
 	}
@@ -73,10 +73,10 @@ void scg_buffer_fill_ch(struct SCGBuffer *buffer, char ch)
 
 void scg_buffer_fill_fg_color(struct SCGBuffer *buffer, enum SCGColorCode fg_color)
 {
-	uint8_t width = buffer->width;
-	uint8_t height = buffer->height;
-	for (uint8_t row = 0; row < height; row++) {
-		for (uint8_t col = 0; col < width; col++) {
+	uint16_t width = buffer->width;
+	uint16_t height = buffer->height;
+	for (uint16_t row = 0; row < height; row++) {
+		for (uint16_t col = 0; col < width; col++) {
 			scg_buffer_set_fg_color(buffer, col, row, fg_color);
 		}
 	}
@@ -84,10 +84,10 @@ void scg_buffer_fill_fg_color(struct SCGBuffer *buffer, enum SCGColorCode fg_col
 
 void scg_buffer_fill_bg_color(struct SCGBuffer *buffer, enum SCGColorCode bg_color)
 {
-	uint8_t width = buffer->width;
-	uint8_t height = buffer->height;
-	for (uint8_t row = 0; row < height; row++) {
-		for (uint8_t col = 0; col < width; col++) {
+	uint16_t width = buffer->width;
+	uint16_t height = buffer->height;
+	for (uint16_t row = 0; row < height; row++) {
+		for (uint16_t col = 0; col < width; col++) {
 			scg_buffer_set_bg_color(buffer, col, row, bg_color);
 		}
 	}
@@ -95,7 +95,7 @@ void scg_buffer_fill_bg_color(struct SCGBuffer *buffer, enum SCGColorCode bg_col
 
 void scg_buffer_make_space(struct SCGBuffer *buffer)
 {
-	for (uint8_t row = 0; row < buffer->height; row++) {
+	for (uint16_t row = 0; row < buffer->height; row++) {
 		printf("\n\x1b[G"); // Add <height> lines to bottom of console
 	}
 }
@@ -109,13 +109,13 @@ void scg_buffer_remove_space(struct SCGBuffer *buffer)
 
 void scg_buffer_print(struct SCGBuffer *buffer)
 {
-	uint8_t width = buffer->width;
-	uint8_t height = buffer->height;
+	uint16_t width = buffer->width;
+	uint16_t height = buffer->height;
 
 	printf("\x1b[G"); // Move to 1st column
 	printf("\x1b[%dA", height); // Move to top of buffer
-	for (uint8_t row = 0; row < height; row++) {
-		for (uint8_t col = 0; col < width; col++) {
+	for (uint16_t row = 0; row < height; row++) {
+		for (uint16_t col = 0; col < width; col++) {
 			scg_print_cell(buffer->cells[row * width + col]);
 		}
 		printf("\x1b[B"); // Move down 1 line
@@ -129,17 +129,17 @@ void scg_buffer_print(struct SCGBuffer *buffer)
 
 static void scg_print_cell(struct SCGCell cell)
 {
-	uint8_t ansi_fg_code = scg_color_code_to_ansi_fg(cell.fg_color);
-	uint8_t ansi_bg_code = scg_color_code_to_ansi_bg(cell.bg_color);
+	uint16_t ansi_fg_code = scg_color_code_to_ansi_fg(cell.fg_color);
+	uint16_t ansi_bg_code = scg_color_code_to_ansi_bg(cell.bg_color);
 	printf("\x1b[%d;%dm%c\x1b[0m", ansi_fg_code, ansi_bg_code, cell.ch); // Print char with specified fg and bg color
 }
 
-static uint8_t scg_color_code_to_ansi_fg(enum SCGColorCode color_code)
+static uint16_t scg_color_code_to_ansi_fg(enum SCGColorCode color_code)
 {
 	return color_code + 39;
 }
 
-static uint8_t scg_color_code_to_ansi_bg(enum SCGColorCode color_code)
+static uint16_t scg_color_code_to_ansi_bg(enum SCGColorCode color_code)
 {
 	return color_code + 49;
 }
